feat: Add parsing for while loops
